refactor(frontend): extract useClusterContext; add Top+4 plate
Phase G foundation for hosting cluster plots on the Analysis canvas:
- useClusterContext(projectUid,imageUids,popType,suffix) — run list + per-run features/cluster IDs/
  members + valid-image resolution + the gating-store drive + highlight→shownPops. Extracted from
  ClusterPlots (now consumes it; behaviour unchanged) so the analysis canvas reuses it (single cluster
  run per board).
- layoutTemplates: add "Top + 4" comic plate (wide banner + a row of four).
